# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prosepect API",
    "description": "Personal productivity API for Prosepect",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/activity": {
      "get": {
        "tags": [
          "synchronization"
        ],
        "operationId": "activity",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityList"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "development_user": []
          }
        ]
      }
    },
...
    "/api/v1/integrations/google": {
      "get": {
        "tags": [
          "synchronization"
        ],
        "operationId": "google_status",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleIntegrationStatus"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "development_user": []
          }
        ]
      },
      "delete": {
        "tags": [
          "synchronization"
        ],
        "operationId": "revoke_google",
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Synchronization"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "development_user": []
          }
        ]
      }
    },
    "/api/v1/integrations/google/calendars/discover": {
      "post": {
        "tags": [
          "synchronization"
        ],
        "operationId": "discover_google_calendars",
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Synchronization"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "development_user": []
          }
        ]
      }
    },
...
    "/api/v1/sync-conflicts": {
      "get": {
        "tags": [
          "synchronization"
        ],
        "operationId": "list_conflicts",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncConflictList"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "development_user": []
          }
        ]
      }
    },
    "/api/v1/sync-conflicts/{conflict_id}/resolve": {
      "post": {
        "tags": [
          "synchronization"
        ],
        "operationId": "resolve_conflict",
        "parameters": [
          {
            "name": "conflict_id",
            "in": "path",
            "description": "Conflict identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveSyncConflictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncConflict"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "development_user": []
          }
        ]
      }
    },
    "/api/v1/synchronizations": {
      "post": {
        "tags": [
          "synchronization"
        ],
        "operationId": "create_synchronization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSynchronizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Synchronization"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "development_user": []
          }
        ]
      }
    },
    "/api/v1/synchronizations/{synchronization_id}": {
      "get": {
        "tags": [
          "synchronization"
        ],
        "operationId": "get_synchronization",
        "parameters": [
          {
            "name": "synchronization_id",
            "in": "path",
            "description": "Synchronization identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Synchronization"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "development_user": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityEntry": {
        "type": "object",
        "required": [
          "kind",
          "message",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "kind": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ActivityList": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityEntry"
            }
          }
        }
      },
...
      "CreateSynchronizationRequest": {
        "type": "object",
        "required": [
          "idempotency_key"
        ],
        "properties": {
          "calendar_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "idempotency_key": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "GoogleIntegrationStatus": {
        "type": "object",
        "required": [
          "connected",
          "scopes"
        ],
        "properties": {
          "connected": {
            "type": "boolean"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ResolveSyncConflictRequest": {
        "type": "object",
        "required": [
          "resolution"
        ],
        "properties": {
          "resolution": {
            "type": "string"
          }
        }
      },
...
      "SyncConflict": {
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "created_at"
        ],
        "properties": {
          "canonical_event_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "resolution": {
            "type": [
              "string",
              "null"
            ]
          },
          "resolved_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "SyncConflictList": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncConflict"
            }
          }
        }
      },
...
      "Synchronization": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "attempt_count",
          "available_at",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "attempt_count": {
            "type": "integer",
            "format": "int32"
          },
          "available_at": {
            "type": "string",
            "format": "date-time"
          },
          "calendar_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string"
          },
          "last_error": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "development_user": {
        "type": "apiKey",
        "in": "header",
        "name": "x-prosepect-user-id",
        "description": "Development-only user identifier."
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "prosepect_session",
        "description": "Secure HTTP-only user session cookie."
      }
    }
  },
...
}